LayoutTests/requestidlecallback/requestidlecallback-in-page-cache-expected.txt

This tests that when requestIdleCallback is not enabled, requestIdleCallback and IdleDeadline are not defined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.persisted is true
PASS logs.length is 0
PASS logs.length is 7
PASS logs.join(", ") is "A1, B1, A2, B2, A3, B3, A4"
PASS successfullyParsed is true

TEST COMPLETE
